+++ tests/test_grammar.py
@@ -31,11 +31,6 @@ class TestCompleteness:
         existing_keywords = set(grammar.SnakeModule.spec)
         self.check_completeness(target_keywords, existing_keywords)
 
-    def test_subworkflow_context_completeness(self):
-        target_keywords = set(parser.Subworkflow.subautomata)
-        existing_keywords = set(grammar.SnakeSubworkflow.spec)
-        self.check_completeness(target_keywords, existing_keywords)
-
     def test_use_rule_context_completeness(self):
         target_keywords = set(parser.UseRule.subautomata)
         existing_keywords = set(grammar.SnakeUseRule.spec)